+++ Python/tdw/add_ons/py_impact.py
@@ -7,7 +7,7 @@ from pkg_resources import resource_filename
 from typing import Dict, Optional, Union, List, Tuple
 import numpy as np
 import scipy.signal as sg
-from scipy.ndimage import gaussian_filter1d
+from scipy.ndimage import gaussian_filter1d, uniform_filter1d
 from pydub import AudioSegment
 from tdw.output_data import OutputData, Rigidbodies, StaticRobot, SegmentationColors, StaticRigidbodies, \
     RobotJointVelocities
@@ -80,13 +80,13 @@ class PyImpact(CollisionManager):
     """
 
     """:class_var
-    50ms of silence. Used for scrapes.
+    100ms of silence. Used for scrapes.
     """
-    SILENCE_50MS: AudioSegment = AudioSegment.silent(duration=50, frame_rate=SAMPLE_RATE)
+    SILENCE_100MS: AudioSegment = AudioSegment.silent(duration=100, frame_rate=SAMPLE_RATE)
     """:class_var
     The maximum velocity allowed for a scrape.
     """
-    SCRAPE_MAX_VELOCITY: float = 5.0
+    SCRAPE_MAX_VELOCITY: float = 1
     """:class_var
     Meters per pixel on the scrape surface.
     """
@@ -716,7 +716,7 @@ class PyImpact(CollisionManager):
             self._scrape_previous_indices[scrape_key] = 0
 
         # Initialize scrape variables; if this is an in=process scrape, these will be replaced bu te stored values.
-        summed_master = AudioSegment.silent(duration=0, frame_rate=SAMPLE_RATE)
+        summed_master = AudioSegment.silent(duration=2000, frame_rate=SAMPLE_RATE)
         scrape_event_count = 0
 
         # Is this a new scrape?
@@ -736,8 +736,8 @@ class PyImpact(CollisionManager):
             self._scrape_start_velocities[scrape_key] = mag
 
         # Map magnitude to gain level -- decrease in velocity = rise in negative dB, i.e. decrease in gain.
-        db2 = 40 * np.log10(mag / PyImpact.SCRAPE_MAX_VELOCITY) - 1
-        db1 = 20 * np.log10(mag / PyImpact.SCRAPE_MAX_VELOCITY) - 10
+        db2 = 40 * np.log10(mag / PyImpact.SCRAPE_MAX_VELOCITY) - 4
+        db1 = 20 * np.log10(mag / PyImpact.SCRAPE_MAX_VELOCITY) - 25
 
         # Get impulse response of the colliding objects. Amp values would normally come from objects.csv.
         # We also get the lowest-frequency IR mode, which we use to set the high-pass filter cutoff below.
@@ -768,9 +768,10 @@ class PyImpact(CollisionManager):
             #   Apply a variable Gaussian average
             #   Calculate the horizontal and vertical forces
             #   Convolve the force with the impulse response
+            scrape_surface = gaussian_filter1d(scrape_surface, 5)
             dsdx = (scrape_surface[1:] - scrape_surface[0:-1]) / PyImpact.SCRAPE_M_PER_PIXEL
             d2sdx2 = (dsdx[1:] - dsdx[0:-1]) / PyImpact.SCRAPE_M_PER_PIXEL
-            rough_ratio = (np.std(scrape_surface) / (9 * 10 ** -4)) ** 0.85
+            rough_ratio = (np.std(scrape_surface) / (3 * 10 ** -4)) ** 1
             r_gain = 20 * np.log10(rough_ratio)
             self.scrape_surface_data[scrape_material] = {"dsdx": dsdx,
                                                          "d2sdx2": d2sdx2,
@@ -791,21 +792,15 @@ class PyImpact(CollisionManager):
         if final_ind > len(self.scrape_surface_data[scrape_material]["surface"]) - 1:
             self._scrape_previous_indices[scrape_key] = 0
             final_ind = num_pts
-            slope_int = np.interp(vect2, vect1, self.scrape_surface_data[scrape_material]["dsdx"][
-                                                self._scrape_previous_indices[scrape_key]:final_ind])
-            curve_int = np.interp(vect2, vect1, self.scrape_surface_data[scrape_material]["d2sdx2"][
-                                                self._scrape_previous_indices[scrape_key]:final_ind])
-        else:
-            slope_int = np.interp(vect2, vect1, self.scrape_surface_data[scrape_material]["dsdx"][
-                                                self._scrape_previous_indices[scrape_key]:final_ind])
-            curve_int = np.interp(vect2, vect1, self.scrape_surface_data[scrape_material]["d2sdx2"][
-                                                self._scrape_previous_indices[scrape_key]:final_ind])
-
+        slope_int = np.interp(vect2, vect1, self.scrape_surface_data[scrape_material]["dsdx"][
+                                            self._scrape_previous_indices[scrape_key]:final_ind])
+        curve_int = np.interp(vect2, vect1, self.scrape_surface_data[scrape_material]["d2sdx2"][
+                                            self._scrape_previous_indices[scrape_key]:final_ind])
         self._scrape_previous_indices[scrape_key] = final_ind
 
-        curve_int_tan = np.tanh(curve_int / (100 * primary_mass))
+        curve_int_tan = np.tanh(curve_int / (1000 * primary_mass))
 
-        d2_section = gaussian_filter1d(curve_int_tan, 6)
+        d2_section = uniform_filter1d(curve_int_tan, 5)
 
         vert_force = d2_section
         hor_force = slope_int
@@ -813,20 +808,8 @@ class PyImpact(CollisionManager):
         t_force2 = vert_force / max(np.abs(vert_force))
         t_force1 = hor_force[:len(vert_force)]
 
-        noise_seg1 = AudioSegment(t_force1.tobytes(),
-                                  frame_rate=SAMPLE_RATE,
-                                  sample_width=SAMPLE_WIDTH,
-                                  channels=CHANNELS)
-        noise_seg2 = AudioSegment(t_force2.tobytes(),
-                                  frame_rate=SAMPLE_RATE,
-                                  sample_width=SAMPLE_WIDTH,
-                                  channels=CHANNELS)
-        # Fade head and tail.
-        noise_seg_fade1 = noise_seg1.fade_in(4).fade_out(4)
-        noise_seg_fade2 = noise_seg2.fade_in(4).fade_out(4)
-        # Convolve the band-pass filtered sound with the impulse response.
-        conv1 = sg.fftconvolve(scraping_ir, noise_seg_fade1.get_array_of_samples())
-        conv2 = sg.fftconvolve(scraping_ir, noise_seg_fade2.get_array_of_samples())
+        conv1 = sg.fftconvolve(scraping_ir, t_force1)
+        conv2 = sg.fftconvolve(scraping_ir, t_force2)
 
         # Again, we need this as an AudioSegment for overlaying with the previous frame's segment.
         # Convert to 16-bit integers for Unity, normalizing to make sure to minimize loss of precision from truncating floating values.
@@ -848,16 +831,14 @@ class PyImpact(CollisionManager):
 
         noise_seg_conv = noise_seg_conv1.overlay(noise_seg_conv2)
         # Apply roughness gain.
-        noise_seg_conv = noise_seg_conv.apply_gain(self.scrape_surface_data[scrape_material]["r_gain"] + 4)
+        noise_seg_conv = noise_seg_conv.apply_gain(self.scrape_surface_data[scrape_material]["r_gain"])
 
-        # Pad the end of master with 50 ms of silence, the start of the current segment with (n * 50ms) of silence, and overlay.
-        padded_current = (PyImpact.SILENCE_50MS * scrape_event_count) + noise_seg_conv
-        summed_master = summed_master + PyImpact.SILENCE_50MS
-        summed_master = summed_master.overlay(padded_current)
-
+        # Pad the end of master with 100ms of silence, the start of the current segment with (n * 100ms) of silence, and overlay.
+        summed_master = summed_master + PyImpact.SILENCE_100MS
+        summed_master = summed_master.overlay(noise_seg_conv, position=100 * scrape_event_count)
         # Extract 100ms "chunk" of sound to send over to Unity.
         start_idx = 100 * scrape_event_count
-        unity_chunk = summed_master[-(len(summed_master) - start_idx):][:100]
+        unity_chunk = summed_master[start_idx:start_idx + 100]
         # Update stored summed waveform.
         self._scrape_summed_masters[scrape_key] = summed_master